move sigmet component into sigmet feature folder
