+++ constants/constants.ts
@@ -1,3 +1,6 @@
 // Initial map location
-export const LATITUDE = 64.5
-export const LONGITUDE = 16.5
+export const LATITUDE = 64.5;
+export const LONGITUDE = 16.5;
+
+export const SIGMET = "rgba(238, 244, 56, 0.44)";
+export const AIRMET = "rgba(0, 183, 255, 0.44)";
